Move TestColumnarArray/Map/Row to presto-main
These tests need access to TypeRegistry, which isn't available in presto-spi.
