+++ presto-main/src/test/java/com/facebook/presto/block/TestColumnarMap.java
@@ -11,22 +11,29 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.spi.block;
+package com.facebook.presto.block;
 
+import com.facebook.presto.spi.block.Block;
+import com.facebook.presto.spi.block.BlockBuilder;
+import com.facebook.presto.spi.block.ColumnarMap;
+import com.facebook.presto.spi.block.DictionaryBlock;
+import com.facebook.presto.spi.block.MapBlockBuilder;
+import com.facebook.presto.spi.block.MethodHandleUtil;
+import com.facebook.presto.spi.block.RunLengthEncodedBlock;
 import io.airlift.slice.Slice;
 import io.airlift.slice.Slices;
 import org.testng.annotations.Test;
 
 import java.lang.invoke.MethodType;
 
+import static com.facebook.presto.block.ColumnarTestUtils.alternatingNullValues;
+import static com.facebook.presto.block.ColumnarTestUtils.assertBlock;
+import static com.facebook.presto.block.ColumnarTestUtils.assertBlockPosition;
+import static com.facebook.presto.block.ColumnarTestUtils.createTestDictionaryBlock;
+import static com.facebook.presto.block.ColumnarTestUtils.createTestDictionaryExpectedValues;
+import static com.facebook.presto.block.ColumnarTestUtils.createTestRleBlock;
+import static com.facebook.presto.block.ColumnarTestUtils.createTestRleExpectedValues;
 import static com.facebook.presto.spi.block.ColumnarMap.toColumnarMap;
-import static com.facebook.presto.spi.block.ColumnarTestUtils.alternatingNullValues;
-import static com.facebook.presto.spi.block.ColumnarTestUtils.assertBlock;
-import static com.facebook.presto.spi.block.ColumnarTestUtils.assertBlockPosition;
-import static com.facebook.presto.spi.block.ColumnarTestUtils.createTestDictionaryBlock;
-import static com.facebook.presto.spi.block.ColumnarTestUtils.createTestDictionaryExpectedValues;
-import static com.facebook.presto.spi.block.ColumnarTestUtils.createTestRleBlock;
-import static com.facebook.presto.spi.block.ColumnarTestUtils.createTestRleExpectedValues;
 import static com.facebook.presto.spi.type.VarcharType.VARCHAR;
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertNotNull;
